first successful run - new model

+++ src/main/java/org/sandag/abm/airport/AirportModelStructure.java
@@ -29,32 +29,32 @@ public final class AirportModelStructure
     public static final int      UPPER_PM            = 29;
     public static final String[] MODEL_PERIOD_LABELS = {"EA", "AM", "MD", "PM", "EV"};
 
-    public static final byte     ACCESS_MODES        = 17;
+    public static final int     ACCESS_MODES        = 17;
 
-    public static final byte     PARK_LOC1           = 1;
-    public static final byte     PARK_LOC2           = 2;
-    public static final byte     PARK_LOC3           = 3;
-    public static final byte     PARK_LOC4		     = 4;
-    public static final byte     PARK_LOC5		     = 5;
-    public static final byte     PARK_ESC		     = 6;
-    public static final byte     RENTAL              = 7;
-    public static final byte     SHUTTLE_VAN         = 8;
-    public static final byte     HOTEL_COURTESY      = 9;
-    public static final byte     RIDEHAILING_LOC1    = 10;
-    public static final byte     RIDEHAILING_LOC2    = 11;
-    public static final byte     TRANSIT             = 12;
-    public static final byte     CURB_LOC1           = 13;
-    public static final byte     CURB_LOC2           = 14;
-    public static final byte     CURB_LOC3           = 15;
-    public static final byte     CURB_LOC4           = 16;
-    public static final byte     CURB_LOC5           = 17;
+    public static final int     PARK_LOC1           = 1;
+    public static final int     PARK_LOC2           = 2;
+    public static final int     PARK_LOC3           = 3;
+    public static final int     PARK_LOC4		     = 4;
+    public static final int     PARK_LOC5		     = 5;
+    public static final int     PARK_ESC		     = 6;
+    public static final int     RENTAL              = 7;
+    public static final int     SHUTTLE_VAN         = 8;
+    public static final int     HOTEL_COURTESY      = 9;
+    public static final int     RIDEHAILING_LOC1    = 10;
+    public static final int     RIDEHAILING_LOC2    = 11;
+    public static final int     TRANSIT             = 12;
+    public static final int     CURB_LOC1           = 13;
+    public static final int     CURB_LOC2           = 14;
+    public static final int     CURB_LOC3           = 15;
+    public static final int     CURB_LOC4           = 16;
+    public static final int     CURB_LOC5           = 17;
     
-    public static final byte     LOS_TYPE        = 4;
+    public static final int     LOS_TYPE        = 4;
     
-    public static final byte     DA        		= 0;
-    public static final byte     SR2        	= 1;
-    public static final byte     SR3        	= 2;
-    public static final byte     Transit        = 3;
+    public static final int     DA        		= 0;
+    public static final int     SR2        	= 1;
+    public static final int     SR3        	= 2;
+    public static final int     Transit        = 3;
 
 
     private AirportModelStructure()
